将循环迭代变量改为 int 类型以优化 iob 函数性能
将 i 的类型由 uint8_t 改为 int 使得编译器可以使用更高效的代码来实现循环。

在很多处理器上，使用 int 类型进行循环计数可以提高性能，
因为这些处理器使用的是 int 类型寄存器，而不是 uint8_t 类型寄存器。
因此，在使用 uint8_t 类型时，编译器可能需要执行一些额外的操作来将其转换为 int 类型，
这可能会降低循环的效率。

此外，使用 int 类型进行循环计数可能还能提高循环展开的效率。
循环展开是编译器将多个循环迭代合并为单个迭代的技术，可以减少循环的开销。
使用 int 类型进行循环计数可以使循环展开的效果更好，
因为编译器可以将更多的迭代合并为单个迭代。

总之，将 i 的类型由 uint8_t 改为 int 会提高函数的性能，
因为这使得编译器可以使用更高效的代码来实现循环，并提高循环展开的效率。

从 7000ms 缩短到 6139ms。

+++ chaosclock3.cpp
@@ -304,8 +304,8 @@ void coutMovelist(vector<uint8_t> &movelist)
 // indexOfBoard
 int8_t iob(uint64_t board, uint8_t c)
 {
-    for (uint8_t i = 0; i < 12; ++i) {
-        if (c == (board >> (i << 2) & 0xf)) {
+    for (int i = 0; i < 12; ++i) {
+        if (c == (board >> ((uint8_t)i << 2) & 0xf)) {
             return i;
         }
     }
